requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.5
choreographer==1.2.1
click==8.3.1
defusedxml==0.7.1
distro==1.9.0
fonttools==4.62.0
fpdf2==2.8.7
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaleido==1.2.0
logistro==2.0.1
MarkupSafe==3.0.3
narwhals==2.18.0
numpy==2.4.3
openai==2.26.0
orjson==3.11.7
packaging==26.0
pandas==2.3.3
pillow==12.1.1
plotly==6.6.0
pluggy==1.6.0
protobuf==6.33.5
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
